feat(#767): moved useParsedProposal logic to useProposalDetails hook
(cherry picked from commit 5ac9618bee3a2c0b10f1207583380deef94bc3de)
