+++ packages/core/src/governance/client/hooks/index.ts
@@ -3,4 +3,3 @@ export * from './useAgreementMutations.web2.rsc';
 export * from './useCreateAgreementOrchestrator';
 export * from './useVote';
 export * from './useProposalDetails.web3.rpc';
-export * from './useParsedProposal';
